Move benchmarks to separate file.
This lets this crate work with the stable Rust build and not
require nightly features to build.
